requirements.txt
Django==5.0.1
djangorestframework==3.14.0
requests==2.31.0
pymongo==4.6.1
mongoengine==0.27.0
django-cors-headers==4.3.1